requirements.txt
# Existing Repository Requirements
langchain==0.0.305
openai==0.28.1
pydantic>=1.10.12
pytest>=7.4.0
sentence-transformers>=2.2.0
transformers>=4.31.0
chromadb==0.4.12
jupyter==1.0.0
tiktoken==0.4.0
python-dotenv